+++ actix-web-codegen/tests/trybuild.rs
@@ -9,6 +9,8 @@ fn compile_macros() {
     t.compile_fail("tests/trybuild/route-missing-method-fail.rs");
     t.compile_fail("tests/trybuild/route-duplicate-method-fail.rs");
     t.compile_fail("tests/trybuild/route-unexpected-method-fail.rs");
+
+    t.pass("tests/trybuild/docstring-ok.rs");
 }
 
 // #[rustversion::not(nightly)]